refactor: 카드 추가 및 수정 화면 관련 네이밍 수정
