+++ app/src/androidTest/java/woowacourse/payments/ui/cardupdate/components/CardHolderNameTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard.components
+package woowacourse.payments.ui.cardupdate.components
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
