+++ app/src/androidTest/java/woowacourse/payments/ui/common/components/PaymentCardTest.kt
@@ -6,8 +6,8 @@ import androidx.compose.ui.test.onNodeWithContentDescription
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.payments.R
+import woowacourse.payments.ui.cardupdate.model.CardCompanyUiModel
 import woowacourse.payments.ui.common.model.CardUiModel
-import woowacourse.payments.ui.newcard.model.CardCompanyUiModel
 
 @Suppress("ktlint:standard:function-naming")
 class PaymentCardTest {
