+++ app/src/main/java/woowacourse/payments/ui/cardupdate/CardUpdateUiState.kt
@@ -1,11 +1,11 @@
-package woowacourse.payments.ui.newcard
+package woowacourse.payments.ui.cardupdate
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.payments.ui.newcard.model.CardCompanyUiModel
+import woowacourse.payments.ui.cardupdate.model.CardCompanyUiModel
 
 @Parcelize
-data class NewCardUiState(
+data class CardUpdateUiState(
     val cardCompany: CardCompanyUiModel? = null,
     val cardNumber: String = "",
     val isCardNumberValid: Boolean = false,
