+++ app/src/main/java/woowacourse/payments/ui/cardupdate/components/CardCompanySelectBottomSheet.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard.components
+package woowacourse.payments.ui.cardupdate.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -12,8 +12,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.domain.CardCompany
-import woowacourse.payments.ui.newcard.model.CardCompanyUiModel
-import woowacourse.payments.ui.newcard.model.toUiModel
+import woowacourse.payments.ui.cardupdate.model.CardCompanyUiModel
+import woowacourse.payments.ui.cardupdate.model.toUiModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
