+++ app/src/main/java/woowacourse/payments/ui/cardupdate/model/CardUpdateType.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard.model
+package woowacourse.payments.ui.cardupdate.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
